Add files via upload
This is a simple Quote Sender program made using SMTP library and Datetime module. 
It sends a quote from one email to another with the help smtplib and random module.
If the day of the week is Monday, the Quote sender will send a random quote from quotes.txt file.